Updating Drop down bar with filled arrow and sizing
Filled the arrow with white, this should make the arrow stand out. It
was transparent before. You should place it on top of the circle which
is already bordered on the drop down bar.
design/drop-down-icon/drop-down-bar-arrow-filled.png

Increased the drop down bar to 640 pixels wide so this should take care
of the completed bar.
design/drop-down-bar-no-arrow.png